smart_home_app/app-gemini/back/requirements.txt
langgraph
langchain
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
# langchain-mcp-adapters # 나중에 실제 라이브러리가 있다면 추가합니다. 